GIX-1187: Improve testing
